root/backend/requirements.txt
Flask-SQLAlchemy
flask
flask-cors
pillow